A module rename to please clippy
