+++ src/tests/integration/compatibility.rs
@@ -1,7 +1,7 @@
 use ::insta::assert_snapshot;
 use ::std::collections::HashMap;
 
-use crate::terminal_pane::PositionAndSize;
+use crate::panes::PositionAndSize;
 use crate::tests::fakes::FakeInputOutput;
 use crate::tests::possible_tty_inputs::Bytes;
 use crate::tests::utils::get_output_frame_snapshots;
